requirements.txt
# Existing Repository Requirements
langchain
openai
pydantic
pytest
sentence-transformers
transformers
chromadb
jupyter
tiktoken
pandas
huggingface-hub
lancedb

# New requirements
streamlit
python-dotenv
numpy
langchain-openai